A Full-Stack, live app, stores slangs belonging to different categories for different countries in the mongoDB cloud NOSql database. It's also linked to an API, for searching up any word and it provides all possible definitions for that word. It's created using nodeJS using Express framework and deployed over Heroku. Hence the stack MongoDB, Express, NodeJS (MEN). The live version can be viewed through the link https://sheltered-island-39681.herokuapp.com/ and a screenshot enclosed gives an overall feel for the app. This app is also named in the york university magazine for its usefulness as it was created for the education purpose for the glendon campus.
A Full-Stack, live app, stores slangs belonging to different categories for different countries in the mongoDB cloud NOSql database. It's also linked to an API, for searching up any word and it provides all possible definitions for that word. It's created using nodeJS using Express framework and deployed over Heroku. Hence the stack MongoDB, Express, NodeJS (MEN). The live version can be viewed through the link https://sheltered-island-39681.herokuapp.com/ and a screenshot enclosed gives an overall feel for the app. This app is also named in the york university magazine for its usefulness as it was created for the education purpose for the glendon campus.